+++ src/froGH/froGH/Properties/Resources.Designer.cs
@@ -113,6 +113,16 @@ namespace froGH.Properties {
         /// <summary>
         ///   Looks up a localized resource of type System.Drawing.Bitmap.
         /// </summary>
+        internal static System.Drawing.Bitmap BitmapFromColorTree_GH {
+            get {
+                object obj = ResourceManager.GetObject("BitmapFromColorTree_GH", resourceCulture);
+                return ((System.Drawing.Bitmap)(obj));
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized resource of type System.Drawing.Bitmap.
+        /// </summary>
         internal static System.Drawing.Bitmap Camera_report_GH {
             get {
                 object obj = ResourceManager.GetObject("Camera_report_GH", resourceCulture);